requirements.txt
presidio_analyzer==2.2.33
presidio_anonymizer==2.2.33
# presidio_image_redactor
loguru
fastapi
uvicorn
python-dotenv
tiktoken==0.3.3
# math